:pencil2: Use local storage for updates (#5)
* Store user in room

* Ensure get in module

* Add kluent and random builder.

* Switch to dto

* Binding is now working !

* Adding data repository tests.

* Adding unit tests for the data repository.

* Cleanup the wrapper.

* Adding more reactive updates and inserts into db.

* Adding create date and update date.

* Adding model mappers.

* Adding glide images.

* Fix formatting.

* More formatting issues.

+++ app/src/main/kotlin/com/coredocker/android/data/network/graphql/PagedFragment.kt
@@ -1,8 +1,6 @@
 package com.coredocker.android.data.network.graphql
 
-import java.math.BigInteger
-
 class PagedFragment<T>(
     val items: List<T>,
-    val count: BigInteger?
+    val count: Int?
 )
